cheep box in component view and pagemodel subclass
