+++ src/Chirp.Web/Pages/Components/CheepBox/Default.cshtml
@@ -1,0 +1,9 @@
+@model Chirp.Web.Pages.Components.CheepablePageModel
+
+<div class="cheepbox">
+    <h3>What's on your mind @(User.Identity.Name)?</h3>
+    <form method="post">
+        <input style="float: left" type="text" asp-for="Text">
+        <input type="submit" value="Share🚜">
+    </form>
+</div>
