+++ src/Chirp.Web/Pages/Public.cshtml
@@ -11,13 +11,7 @@
 
     @if (User.Identity.IsAuthenticated)
     {
-        <div class="cheepbox">
-            <h3>What's on your mind @(User.Identity.Name)?</h3>
-            <form method="post">
-                <input style="float: left" type="text" asp-for="Text">
-                <input type="submit" value="Share🚜">
-            </form>
-        </div>
+        @await Component.InvokeAsync("CheepBox" , Model);
     }
 
     @if (Model.Cheeps != null && Model.Cheeps.Any())
