+++ src/Chirp.Web/Pages/UserTimeline.cshtml
@@ -12,64 +12,59 @@
     @inject ICheepRepository CheepRepository
     @if (User.Identity != null && User.Identity.IsAuthenticated)
     {
-        <div class="cheepbox">
-            <h3>What's on your mind @(User.Identity.Name)?</h3>
-            <form method="post">
-                <input style="float: left" type="text" asp-for="Text">
-                <input type="submit" value="Share🚜">
-            </form>
-        </div>
+        @await Component.InvokeAsync("CheepBox", Model)
     }
 
     @if (Model.Cheeps != null && Model.Cheeps.Any())
     {
-        <ul id=" messagelist" class="cheeps">
+                <ul id=" messagelist" class="cheeps">
             @foreach (var cheep in Model.Cheeps)
             {
-                <li>
-                    <p>
-                        <strong>
-                            <a href="/@cheep.Author">@cheep.Author</a>
-                        </strong>
+                                <li>
+                                    <p>
+                                        <strong>
+                                            <a href="/@cheep.Author">@cheep.Author</a>
+                                        </strong>
                         @cheep.Message
-                        <small>&mdash; @cheep.Timestamp</small>
-                    </p>
+                                        <small>&mdash; @cheep.Timestamp</small>
+                                    </p>
 
                     @if (User.Identity != null && User.Identity.IsAuthenticated && cheep.Author != User.Identity.Name)
                     {
                         @if ((Model.FollowedAuthors ?? new List<Guid>()).Contains(cheep.AuthorId))
                         {
-                            <form method="post">
-                                <input type="hidden" name="followName" value="@cheep.Author">
-                                <button asp-page-handler="Unfollow">Unfollow</button>
-                            </form>
+                                                            <form method="post">
+                                                                <input type="hidden" name="followName" value="@cheep.Author">
+                                                                <button asp-page-handler="Unfollow">Unfollow</button>
+                                                            </form>
                         }
                         else
                         {
-                            <form method="post">
-                                <input type="hidden" name="followName" value="@cheep.Author">
-                                <button asp-page-handler="Follow">Follow</button>
-                            </form>
+                                                            <form method="post">
+                                                                <input type="hidden" name="followName" value="@cheep.Author">
+                                                                <button asp-page-handler="Follow">Follow</button>
+                                                            </form>
                         }
                     }
-                </li>
+                                </li>
             }
-        </ul>
-        <div class="navigation">
+                </ul>
+                <div class="navigation">
             @if (Model.PageNumber > 1)
             {
-                <a href="/@routeName/?page=@(Model.PageNumber - 1)"> Previous </a>
+                                <a href="/@routeName/?page=@(Model.PageNumber - 1)"> Previous </a>
             }
-            | <a href="/@routeName/?page=@(Model.PageNumber - 1)"> @Model.PageNumber</a> |
-            @if(Model.HasNextPage) {
-                <a href="/@routeName/?page=@(Model.PageNumber + 1)"> Next </a>
+                    | <a href="/@routeName/?page=@(Model.PageNumber - 1)"> @Model.PageNumber</a> |
+            @if (Model.HasNextPage)
+            {
+                                <a href="/@routeName/?page=@(Model.PageNumber + 1)"> Next </a>
             }
-            
-        </div>
+
+                </div>
     }
 
     else
     {
-        <em>There are no cheeps so far.</em>
+                <em>There are no cheeps so far.</em>
     }
 </div>
